testes 2

+++ backend/src/main/java/com/devsuperior/dscatalog/services/ProductService.java
@@ -71,6 +71,9 @@ public class ProductService {
 
     @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
+        if(!repository.existsById(id)){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
         try {
             repository.deleteById(id);
         }
